Guess The Number
Generate a random number from a to b. You and your friend have to guess a number between two number a and b. a and b are inputs taken form the user. Your friend is players 1 and plays first. He will actual number or less than the actual number. Log the number of trials it look your friend to arrive at the number. You play the same game and then the person with minimum number of trails wins